+++ ai/src/main/kotlin/com/oneliang/ktx/frame/ai/cnn/layer/FullyConnectedLayer.kt
@@ -7,19 +7,19 @@ import com.oneliang.ktx.frame.ai.dnn.layer.LossLayer
 
 open class FullyConnectedLayer<IN : Any, OUT : Any, LOSS : Any>(
     val neuronCount: Int,//32
-    private val forwardImpl: ((layer: FullyConnectedLayer<IN, OUT, LOSS>, dataId: Long, inputNeuron: IN, y: Double, training: Boolean) -> OUT)? = null,
-    private val backwardImpl: ((layer: FullyConnectedLayer<IN, OUT, LOSS>, dataId: Long, inputNeuron: IN, y: Double) -> Unit)? = null,
+    private val forwardImpl: ((layer: FullyConnectedLayer<IN, OUT, LOSS>, dataId: Long, inputNeuron: IN, y: Float, training: Boolean) -> OUT)? = null,
+    private val backwardImpl: ((layer: FullyConnectedLayer<IN, OUT, LOSS>, dataId: Long, inputNeuron: IN, y: Float) -> Unit)? = null,
     private val forwardResetImpl: ((layer: FullyConnectedLayer<IN, OUT, LOSS>, dataId: Long) -> Unit)? = null,
-    private val updateImpl: ((layer: FullyConnectedLayer<IN, OUT, LOSS>, epoch: Int, printPeriod: Int, totalDataSize: Long, learningRate: Double) -> Unit)? = null,
+    private val updateImpl: ((layer: FullyConnectedLayer<IN, OUT, LOSS>, epoch: Int, printPeriod: Int, totalDataSize: Long, learningRate: Float) -> Unit)? = null,
     private val initializeLayerModelDataImpl: ((layer: FullyConnectedLayer<IN, OUT, LOSS>, data: String) -> Unit) = { _, _ -> },
     private val saveLayerModelDataImpl: ((layer: FullyConnectedLayer<IN, OUT, LOSS>) -> String) = { Constants.String.BLANK }
 ) : LossLayer<IN, OUT, LOSS>() {
 
-    override fun forwardImpl(dataId: Long, inputNeuron: IN, y: Double, training: Boolean): OUT {
+    override fun forwardImpl(dataId: Long, inputNeuron: IN, y: Float, training: Boolean): OUT {
         return this.forwardImpl?.invoke(this, dataId, inputNeuron, y, training) ?: outputNullError()
     }
 
-    override fun backwardImpl(dataId: Long, inputNeuron: IN, y: Double) {
+    override fun backwardImpl(dataId: Long, inputNeuron: IN, y: Float) {
         this.backwardImpl?.invoke(this, dataId, inputNeuron, y)
     }
 
@@ -27,7 +27,7 @@ open class FullyConnectedLayer<IN : Any, OUT : Any, LOSS : Any>(
         this.forwardResetImpl?.invoke(this, dataId)
     }
 
-    override fun updateImpl(epoch: Int, printPeriod: Int, totalDataSize: Long, learningRate: Double) {
+    override fun updateImpl(epoch: Int, printPeriod: Int, totalDataSize: Long, learningRate: Float) {
         this.updateImpl?.invoke(this, epoch, printPeriod, totalDataSize, learningRate)
     }
